+++ FSMPlayers/AllEnemy.py
@@ -1,10 +1,12 @@
 import arcade
 import math
-from util.constants import * 
+from util import constants
+from util.constants import ARROW_SPEED, MOVEMENT_SPEED
+
 
 class Enemy(arcade.Sprite):
     def throwfire(self):
-        fireball = Fireball("images/fire.png", .1)
+        fireball = constants.Fireball("images/fire.png", .1)
         fireball.center_x = self.center_x
         fireball.center_y = self.center_y
         fireball.start_x = self.center_x # for tracking 
@@ -16,7 +18,7 @@ class Enemy(arcade.Sprite):
         self.fireball_list.append(fireball)
 
     def shootarrow(self):
-        arrow = Arrow("images/arrow.png",.1)
+        arrow = constants.Arrow("images/arrow.png",.1)
         arrow.center_x = self.center_x
         arrow.center_y = self.center_y
         arrow.angle = self.angle-90
@@ -30,7 +32,7 @@ class Enemy(arcade.Sprite):
         self.shield +=1
 
     def shortattack(self):
-        knife = Knife("images/knife.png",.1)
+        knife = constants.Knife("images/knife.png",.1)
         knife.center_x = self.center_x
         knife.center_y = self.center_y
         knife.angle = self.angle-180
