+++ GENN/GENNFunctions.py
@@ -1,6 +1,7 @@
 import random 
 import numpy as np
-from util.constants import * 
+from util.neural_net import *
+from util import constants
 import csv
 import ast
 
@@ -20,12 +21,12 @@ def createNets(conCurrentGame, adaptive = False):
                 nodeWeights = np.random.rand(1,inputsNum,totalNodes[0]).tolist()[0]
             else:
                 nodeWeights = np.random.rand(1,totalNodes[j-1],totalNodes[j]).tolist()[0]
-            layers.append(Layer(nodeWeights))
-        layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-        layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-        layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-        layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-        layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+            layers.append(constants.Layer(nodeWeights))
+        layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+        layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+        layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+        layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+        layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
         if adaptive == True: #JTW modified to optionally allow creating adaptive network which adjusts its own weights
             nets.append(AdaptiveNetwork(layers))
         else:
@@ -48,12 +49,12 @@ def createNet(specificLayers = None,specificNodes = None, adaptive = False):
             nodeWeights = np.random.rand(1,inputsNum,totalNodes[0]).tolist()[0]
         else:
             nodeWeights = np.random.rand(1,totalNodes[j-1],totalNodes[j]).tolist()[0]
-        layers.append(Layer(nodeWeights))
-    layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-    layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-    layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-    layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
-    layers.append(Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+        layers.append(constants.Layer(nodeWeights))
+    layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+    layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+    layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+    layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
+    layers.append(constants.Layer(np.random.rand(1,totalNodes[len(totalNodes)-1],1).tolist()[0]))
     if adaptive == True:  #JTW modified to optionally allow creating adaptive network which adjusts its own weights
         return AdaptiveNetwork(layers)
     else:
@@ -135,7 +136,7 @@ def readNets(nets, adaptive = False):
             temp = []
             for j in row:
                 temp.append(ast.literal_eval(j))
-            layers.append(Layer(temp))
+            layers.append(constants.Layer(temp))
         if adaptive:
             return[AdaptiveNetwork(layers)] * nets 
         else:
