+++ SimulatedGame.py
@@ -8,9 +8,12 @@ from FSMPlayers.RangePlayer import *
 from FSMPlayers.MidRangePlayer import *
 from FSMPlayers.ShortRangePlayer import *
 from FSMPlayers.AllEnemy import *
-from FSMPlayers.HumanPlayer import *
 from util.inputFunctions import *
 from DynamicController.DynamicController import *
+
+from util import constants
+from util.constants import KNIGHT_IMAGE, FIREBALL_DAMAGE, ARROW_DAMAGE, KNIFE_DAMAGE
+
 
 class SimulatedGame(arcade.Window):
     """ Main application class. """
