+++ sim.py
@@ -14,10 +14,21 @@ from FSMPlayers.HumanPlayer import *
 from util.inputFunctions import *
 from DynamicController.DynamicControllerSim import *
 from GENN.GENN import * 
-from util.constants import *
+from util.neural_net import *
 import multiprocessing
 import numpy as np 
 import json
+
+from util.constants import RANDOM_SEED, \
+    SCREEN_HEIGHT, \
+    SCREEN_WIDTH, \
+    PLAYER_HEALTH, \
+    KNIGHT_IMAGE, \
+    ARROW_DAMAGE, \
+    FIREBALL_DAMAGE, \
+    KNIFE_DAMAGE, ARROW_SPEED
+
+from util import constants
 
 random.seed(RANDOM_SEED)
 
@@ -334,22 +345,22 @@ class Game:
     # There are 2 functions for each so it is easy to not interfere 
     # with each other
     def arrow1(self):
-        self.arw = ArrowSimulated(self.player1.center_x,self.player1.center_y,ARROW_SPEED,self.player1.box)
+        self.arw = constants.ArrowSimulated(self.player1.center_x,self.player1.center_y,ARROW_SPEED,self.player1.box)
         self.player1.arrow_list.append(self.arw)
         
     def arrow2(self):
 
-        self.arw = ArrowSimulated(self.player2.center_x,self.player2.center_y,ARROW_SPEED,self.player2.box)
+        self.arw = constants.ArrowSimulated(self.player2.center_x,self.player2.center_y,ARROW_SPEED,self.player2.box)
         self.player2.arrow_list.append(self.arw)
         
     def fire1(self):
 
-        self.fireball = FireballSimulated(self.player1.center_x,self.player1.center_y,ARROW_SPEED,self.player1.box)  
+        self.fireball = constants.FireballSimulated(self.player1.center_x,self.player1.center_y,ARROW_SPEED,self.player1.box)
         self.player1.fireball_list.append(self.fireball)
         
     def fire2(self):
 
-        self.fireball = FireballSimulated(self.player2.center_x,self.player2.center_y,ARROW_SPEED,self.player2.box)  
+        self.fireball = constants.FireballSimulated(self.player2.center_x,self.player2.center_y,ARROW_SPEED,self.player2.box)
         self.player2.fireball_list.append(self.fireball)
     
     def equip_shield1(self):
